+++ stories/components/data-table/docs/data-table.docs.stories.tsx
@@ -71,6 +71,16 @@ export const CheckboxInTable: StoryObj<typeof CheckboxInTableProps> = {
       control: { type: 'number' },
     },
     loading: { control: 'boolean' },
+    sideIndents: {
+      control: 'select',
+      options: [undefined, 'wide'],
+      defaultValue: undefined,
+    },
+    compact: {
+      control: 'radio',
+      options: [undefined, true, false],
+      defaultValue: undefined,
+    },
   },
   args: CheckboxInTableProps,
 };